// src/app/auth/interfaces/auth-response.ts
export interface AuthResponse {
    email_user:      string;
    distid:          string;
    role:            number;
    token:           string;
    additional_data: AdditionalData;
}

export interface AdditionalData {
    nombre:          string;
    apellido:        string;
    cedula:          string;
    fechaNacimiento: Date;
}